GUIRAmark1/integrations/guira_core/vision/probes/yolo_probe/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
Pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
ultralytics>=8.3.0
pydantic>=2.0.0
